+++ services/payments-core/src/main/java/br/com/arinelli/pay/payments/outbox/OutboxEventRepository.java
@@ -1,0 +1,6 @@
+package br.com.arinelli.pay.payments.outbox;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OutboxEventRepository extends JpaRepository<OutboxEvent, Long> {
+}