+++ services/payments-core/src/main/java/br/com/arinelli/pay/payments/webhooks/WebhookEventRepository.java
@@ -1,0 +1,6 @@
+package br.com.arinelli.pay.payments.webhooks;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface WebhookEventRepository extends JpaRepository<WebhookEvent, Long> {
+}